# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E.Practice API",
    "description": "Modulo case management per studi professionali — backend FastAPI V0 (in-memory). Differenziatore: vista albero della pratica con eventi ad-hoc agganciati alle fasi.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/templates/{category_id}": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "List Templates For Category",
        "description": "Lista fasi del template per la categoria, ordinate per order_index.",
        "operationId": "list_templates_for_category_api_templates__category_id__get",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PhaseTemplate"
                  },
                  "title": "Response List Templates For Category Api Templates  Category Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/templates/category/{category_id}/preview": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "Template Preview",
        "description": "Restituisce le fasi del template con date pianificate cumulative\na partire da `apertura`, e calcola la scadenza finale.\n\nUsato dal modal Nuova Pratica:\n- Step 2 mostra la scadenza calcolata\n- Step 3 mostra le N fasi con durata e anteprima date",
        "operationId": "template_preview_api_templates_category__category_id__preview_get",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          },
          {
            "name": "apertura",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data apertura ipotetica (default oggi)",
              "title": "Apertura"
            },
            "description": "Data apertura ipotetica (default oggi)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PhaseTemplate": {
        "properties": {
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Days"
          },
          "default_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Role"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "order_index",
          "name",
          "id"
        ],
        "title": "PhaseTemplate"
      },
...
      "TemplatePhasePreview": {
        "properties": {
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_days": {
            "type": "integer",
            "title": "Duration Days"
          },
          "planned_start": {
            "type": "string",
            "format": "date",
            "title": "Planned Start"
          },
          "planned_end": {
            "type": "string",
            "format": "date",
            "title": "Planned End"
          }
        },
        "type": "object",
        "required": [
          "order_index",
          "name",
          "duration_days",
          "planned_start",
          "planned_end"
        ],
        "title": "TemplatePhasePreview"
      },
      "TemplatePreview": {
        "properties": {
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "category_name": {
            "type": "string",
            "title": "Category Name"
          },
          "apertura": {
            "type": "string",
            "format": "date",
            "title": "Apertura"
          },
          "scadenza_calcolata": {
            "type": "string",
            "format": "date",
            "title": "Scadenza Calcolata"
          },
          "total_duration_days": {
            "type": "integer",
            "title": "Total Duration Days",
            "description": "Somma di duration_days delle fasi (giorni cumulativi)."
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/TemplatePhasePreview"
            },
            "type": "array",
            "title": "Phases"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "category_name",
          "apertura",
          "scadenza_calcolata",
          "total_duration_days",
          "phases"
        ],
        "title": "TemplatePreview",
        "description": "Anteprima fasi + scadenza calcolata, per modal Nuova Pratica step 2/3."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}